BD-1/BD1_Simulator_v2_full/requirements.txt
pyttsx3
pygame
PyYAML
sounddevice
vosk
ollama
tkinter
